requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
faiss-cpu>=1.8.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
streamlit>=1.38.0
langsmith>=0.1.100
python-dotenv>=1.0.0
pypdf>=4.0.0
pydantic>=2.7.0
pandas>=2.2.0
numpy>=1.26.0
tiktoken>=0.7.0
